+++ custom_components/tcx_direct/const.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 DOMAIN = "tcx_direct"
 NAME = "Jandy TCX Direct"
-VERSION = "0.2.4"
+VERSION = "0.2.5"
 
 ATTR_RPM = "rpm"
 SERVICE_START_PUMP_AT_SPEED = "start_pump_at_speed"